publisher/requirements.txt
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiokafka>=0.10.0
python-dotenv>=1.0.0
